+++ app/src/main/java/com/infinity/omos/ui/searchtab/ArtistFragment.kt
@@ -33,6 +33,7 @@ class ArtistFragment : Fragment() {
     lateinit var broadcastReceiver: BroadcastReceiver
 
     private var page = 0
+    private var isLoading = false
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -82,6 +83,7 @@ class ArtistFragment : Fragment() {
 
                 mAdapter.setRecord(it)
                 mAdapter.submitList(it)
+                isLoading = false
             }
         })
 
@@ -118,7 +120,8 @@ class ArtistFragment : Fragment() {
                 val itemTotalCount = recyclerView.adapter!!.itemCount-1
 
                 // 스크롤이 끝에 도달했는지 확인
-                if (!binding.recyclerView.canScrollVertically(1) && lastVisibleItemPosition == itemTotalCount) {
+                if (!binding.recyclerView.canScrollVertically(1) && lastVisibleItemPosition == itemTotalCount && !isLoading) {
+                    isLoading = true
                     mAdapter.deleteLoading()
                     viewModel.loadMoreArtist(MainActivity.keyword, 20, ++page*20)
                 }
