+++ app/src/main/java/com/infinity/omos/ui/searchtab/MusicFragment.kt
@@ -31,6 +31,7 @@ class MusicFragment : Fragment() {
     lateinit var broadcastReceiver: BroadcastReceiver
 
     private var page = 0
+    private var isLoading = false
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -69,6 +70,7 @@ class MusicFragment : Fragment() {
 
                 mAdapter.setRecord(it)
                 mAdapter.submitList(it)
+                isLoading = false
             }
         })
 
@@ -105,7 +107,8 @@ class MusicFragment : Fragment() {
                 val itemTotalCount = recyclerView.adapter!!.itemCount-1
 
                 // 스크롤이 끝에 도달했는지 확인
-                if (!binding.recyclerView.canScrollVertically(1) && lastVisibleItemPosition == itemTotalCount) {
+                if (!binding.recyclerView.canScrollVertically(1) && lastVisibleItemPosition == itemTotalCount && !isLoading) {
+                    isLoading = true
                     mAdapter.deleteLoading()
                     viewModel.loadMoreMusic(MainActivity.keyword, 20, ++page*20)
                 }
